refactor(render): mechanical split of render/jax/jax_renderer.py into sibling generators
MAJ-04 module 3/6: render/jax/jax_renderer.py 2200 -> 170 lines (facade).

- jax_spec_extract.py: GNN matrix extraction/validation/repair (leaf)
- jax_factorized_generator.py: Kronecker-factorized (MAJ-02) script generation
- jax_model_generator.py: general JAX model script + EFE convention constants
- jax_pomdp_generator.py: POMDP solver script generation
- jax_combined_generator.py: combined hierarchical/multi-agent/continuous script
- jax_renderer.py: facade re-exporting all 22 moved names; the four public
  render_* routing entrypoints retained verbatim

DAG: jax_spec_extract <- generators <- facade. 2190 lines verified
byte-identical against the pre-split file; only import headers differ.
Package __init__, route tables (render/processor.py, pomdp_processor.py,
framework_registry.py), and pipeline_validator text-scan consumer all
unchanged (the latter is guarded by a pre-rename path and never fires).

Verified: ruff 0 violations; mypy 0 errors; facade contract 27 names +
string-dispatch package surface intact; targeted pytest
test_jax_renderer + test_jax_factorized_pipeline + test_continuous_renderers
+ test_pipeline_step3_step11_step12_integration: 58 passed, 1 skipped (env).
