+++ src/gnn/render/jax/jax_combined_generator.py
@@ -1,0 +1,213 @@
+#!/usr/bin/env python3
+"""
+Combined hierarchical/multi-agent/continuous JAX script generation for GNN Step 11.
+
+Extracted from ``render.jax.jax_renderer``.
+"""
+
+from typing import (
+    Any,
+    Dict,
+    Optional,
+)
+
+from .jax_spec_extract import _jax_model_name
+
+
+def _generate_jax_combined_code(
+    gnn_spec: Dict[str, Any], options: Optional[Dict[str, Any]]
+) -> str:
+    """Generate JAX code for hierarchical/multi-agent/continuous models."""
+
+    model_name = _jax_model_name(gnn_spec, "CombinedModel")
+
+    code = f'''"""
+JAX Combined Model Generated from GNN Specification: {model_name}
+
+This implements a combined JAX model supporting hierarchical, multi-agent, and continuous extensions.
+Uses advanced JAX features including distributed computing and mixed precision.
+
+@Web: https://github.com/google/jax
+@Web: https://optax.readthedocs.io
+@Web: https://flax.readthedocs.io
+"""
+
+import jax
+import jax.numpy as jnp
+import flax.linen as nn
+from functools import partial
+from jax import jit, vmap, pmap
+from typing import Dict, Any, Optional, Tuple, List
+import optax
+
+class {model_name}Combined(nn.Module):
+    """
+    Combined JAX model supporting hierarchical, multi-agent, and continuous extensions.
+    """
+    
+    # Model configuration
+    num_agents: int = 1
+    num_hierarchical_levels: int = 1
+    continuous_dimensions: int = 0
+    use_mixed_precision: bool = True
+    
+    def setup(self):
+        """Initialize model parameters."""
+        # Hierarchical parameters
+        self.hierarchical_weights = []
+        for level in range(self.num_hierarchical_levels):
+            level_weight = self.param(
+                f"hierarchical_{{level}}",
+                nn.initializers.normal(0.1),
+                (self.num_agents, self.num_agents),
+            )
+            self.hierarchical_weights.append(level_weight)
+        
+        # Multi-agent communication parameters
+        if self.num_agents > 1:
+            self.communication_matrix = self.param('communication_matrix',
+                                                 nn.initializers.orthogonal(),
+                                                 (self.num_agents, self.num_agents))
+        
+        # Continuous state parameters
+        if self.continuous_dimensions > 0:
+            self.continuous_encoder = nn.Dense(self.continuous_dimensions)
+            self.continuous_decoder = nn.Dense(self.continuous_dimensions)
+    
+    @nn.compact
+    def __call__(self, inputs: Dict[str, jnp.ndarray], training: bool = False) -> Dict[str, jnp.ndarray]:
+        """
+        Forward pass of the combined model.
+        
+        Args:
+            inputs: Dictionary containing:
+                - 'agent_states': [num_agents, state_dim] - Individual agent states
+                - 'hierarchical_context': [num_levels, context_dim] - Hierarchical context
+                - 'continuous_inputs': [batch_size, continuous_dim] - Continuous inputs
+                - 'communication_mask': [num_agents, num_agents] - Communication topology
+            training: Whether in training mode
+            
+        Returns:
+            Dictionary containing:
+                - 'agent_outputs': [num_agents, output_dim] - Individual agent outputs
+                - 'hierarchical_outputs': [num_levels, output_dim] - Hierarchical outputs
+                - 'continuous_outputs': [batch_size, continuous_dim] - Continuous outputs
+                - 'communication_outputs': [num_agents, num_agents] - Communication outputs
+        """
+        # Extract inputs with defaults
+        agent_states = inputs.get('agent_states', jnp.zeros((self.num_agents, 1)))
+        hierarchical_context = inputs.get('hierarchical_context', jnp.zeros((self.num_hierarchical_levels, 1)))
+        continuous_inputs = inputs.get('continuous_inputs', jnp.zeros((1, self.continuous_dimensions)))
+        communication_mask = inputs.get('communication_mask', jnp.eye(self.num_agents))
+        
+        # The following line generates 'outputs = {{{{}}}}' in the output code
+        outputs = {{}}
+        
+        # 1. Multi-agent processing
+        if self.num_agents > 1:
+            # Apply communication matrix with mask
+            communication_weights = self.communication_matrix * communication_mask
+            
+            # Use vmap to process multiple agents efficiently
+            # jnp.dot(communication_weights, agent_states) is already vectorized, 
+            # but we can explicitly define per-agent processing for complex extensions
+            @jax.vmap
+            def process_agent(weights_row, states):
+                return jnp.dot(weights_row, states)
+                
+            agent_outputs = process_agent(communication_weights, agent_states)
+            outputs['agent_outputs'] = agent_outputs
+            outputs['communication_outputs'] = communication_weights
+        
+        # 2. Hierarchical processing
+        if self.num_hierarchical_levels > 1:
+            hierarchical_outputs = []
+            for level in range(self.num_hierarchical_levels):
+                if level < len(self.hierarchical_weights):
+                    level_output = jnp.dot(self.hierarchical_weights[level], 
+                                         hierarchical_context[level])
+                    hierarchical_outputs.append(level_output)
+                else:
+                    # Default processing for missing levels
+                    hierarchical_outputs.append(hierarchical_context[level])
+            
+            outputs['hierarchical_outputs'] = jnp.stack(hierarchical_outputs)
+        
+        # 3. Continuous processing
+        if self.continuous_dimensions > 0:
+            # Encode continuous inputs
+            encoded = self.continuous_encoder(continuous_inputs)
+            
+            # Apply activation and processing
+            processed = jax.nn.relu(encoded)
+            
+            # Decode back to continuous space
+            decoded = self.continuous_decoder(processed)
+            
+            outputs['continuous_outputs'] = decoded
+        
+        # 4. Combined output (if multiple components exist)
+        if len(outputs) > 1:
+            # Combine different outputs using weighted sum
+            combined_components = []
+            weights = []
+            
+            if 'agent_outputs' in outputs:
+                combined_components.append(outputs['agent_outputs'].flatten())
+                weights.append(1.0)
+            
+            if 'hierarchical_outputs' in outputs:
+                combined_components.append(outputs['hierarchical_outputs'].flatten())
+                weights.append(0.5)
+            
+            if 'continuous_outputs' in outputs:
+                combined_components.append(outputs['continuous_outputs'].flatten())
+                weights.append(0.3)
+            
+            # Normalize weights
+            weights = jnp.array(weights) / jnp.sum(weights)
+            
+            # Weighted combination
+            combined_output = jnp.zeros_like(combined_components[0])
+            for component, weight in zip(combined_components, weights):
+                # Pad or truncate to match size
+                if len(component) > len(combined_output):
+                    component = component[:len(combined_output)]
+                elif len(component) < len(combined_output):
+                    padding = jnp.zeros(len(combined_output) - len(component))
+                    component = jnp.concatenate([component, padding])
+                
+                combined_output += weight * component
+            
+            outputs['combined_output'] = combined_output
+        
+        return outputs
+    
+    def get_parameters(self) -> Dict[str, jnp.ndarray]:
+        """Get all model parameters."""
+        return {
+        "hierarchical_weights": self.hierarchical_weights,
+            'communication_matrix': getattr(self, 'communication_matrix', None),
+            'continuous_encoder': self.continuous_encoder.variables if hasattr(self, 'continuous_encoder') else None,
+            'continuous_decoder': self.continuous_decoder.variables if hasattr(self, 'continuous_decoder') else None
+        }
+    
+    def update_parameters(self, new_params: Dict[str, jnp.ndarray]):
+        """Update model parameters (for training)."""
+        # Implementation for parameter updating during training
+        # This would typically involve gradient-based updates
+        # For now, we'll provide a basic structure
+        updated_params = {{}}
+        
+        for param_name, new_value in new_params.items():
+            if param_name in self.get_parameters():
+                updated_params[param_name] = new_value
+        
+        return updated_params
+
+if __name__ == "__main__":
+    print(f"Combined model {model_name} created successfully!")
+    print("This model supports hierarchical, multi-agent, and continuous extensions.")
+'''
+
+    return code
