+++ src/gnn/render/jax/jax_renderer.py
@@ -136,2065 +136,35 @@ def render_gnn_to_jax_combined(
     return _render_to_path(
         _generate_jax_combined_code, "JAX combined", gnn_spec, output_path, options
     )
-
-
-# --- Kronecker-factorized (MAJ-02) helpers ---
-
-_FACTOR_MATRIX_RE = re.compile(r"^([ABCD])_f(\d+)$")
-
-
-def _is_factorized_spec(gnn_spec: Dict[str, Any]) -> bool:
-    """True when the spec declares per-factor ``A_fN``/``B_fN`` matrices."""
-    matrices = (gnn_spec.get("structured_pomdp") or {}).get("matrices") or {}
-    a_factor_keys = [
-        key
-        for key in matrices
-        if _FACTOR_MATRIX_RE.match(str(key)) and str(key).startswith("A_f")
-    ]
-    b_factor_keys = [
-        key
-        for key in matrices
-        if _FACTOR_MATRIX_RE.match(str(key)) and str(key).startswith("B_f")
-    ]
-    return len(a_factor_keys) >= 2 and len(b_factor_keys) >= 2
-
-
-def _factor_matrix_groups(
-    gnn_spec: Dict[str, Any],
-) -> Dict[int, Dict[str, Any]]:
-    """Group per-factor A/B/C/D matrices keyed by factor index."""
-    matrices = (gnn_spec.get("structured_pomdp") or {}).get("matrices") or {}
-    groups: Dict[int, Dict[str, Any]] = {}
-    for key, value in matrices.items():
-        match = _FACTOR_MATRIX_RE.match(str(key))
-        if not match:
-            continue
-        letter, index = match.group(1), int(match.group(2))
-        groups.setdefault(index, {})[letter] = value
-    return {
-        index: group
-        for index, group in sorted(groups.items())
-        if set(group) >= {"A", "B", "C", "D"}
-    }
-
-
-def _canonicalise_factor_b(value: Any, num_actions: int) -> np.ndarray:
-    """Canonicalise a per-factor B tensor to ``(next, previous, action)``.
-
-    Mirrors ``POMDPRenderProcessor._canonicalise_factored_B``: action-major
-    raw tensors (the generator layout) are transposed and each action slice
-    is column-normalised.
-    """
-    raw = np.asarray(value, dtype=np.float64)
-    if raw.ndim == 2:
-        tensor = raw[:, :, np.newaxis]
-    elif raw.ndim == 3 and raw.shape[0] == num_actions and raw.shape[1] == raw.shape[2]:
-        tensor = raw.transpose(2, 1, 0)
-    elif raw.ndim == 3 and raw.shape[0] == 1 and raw.shape[1] == raw.shape[2]:
-        tensor = raw.transpose(2, 1, 0)
-    elif raw.ndim == 3 and raw.shape[0] == raw.shape[1]:
-        tensor = raw
-    elif raw.ndim == 3:
-        tensor = raw
-    else:
-        raise ValueError(f"B factor must be 2D or 3D, got shape {raw.shape}")
-    tensor = tensor.astype(np.float64, copy=True)
-    for action in range(tensor.shape[2]):
-        column_sums = tensor[:, :, action].sum(axis=0)
-        if np.any(~np.isfinite(column_sums)) or np.any(column_sums <= 0):
-            raise ValueError(
-                f"B factor column has invalid probability mass for action {action}"
-            )
-        tensor[:, :, action] /= column_sums[np.newaxis, :]
-    return tensor
-
-
-def _generate_jax_factorized_code(
-    gnn_spec: Dict[str, Any], options: Optional[Dict[str, Any]]
-) -> str:
-    """Generate a standalone sparse Kronecker-factorized JAX script."""
-    groups = _factor_matrix_groups(gnn_spec)
-    if not groups:
-        raise ValueError("No complete per-factor A/B/C/D groups found")
-    model_name = _jax_model_name(gnn_spec, "GNNModel")
-    safe_model_name = re.sub(r"[^A-Za-z0-9_.-]+", "_", model_name).strip("._")
-
-    a_lists: List[Any] = []
-    b_lists: List[Any] = []
-    c_lists: List[Any] = []
-    d_lists: List[Any] = []
-    for index in sorted(groups):
-        group = groups[index]
-        b_canonical = _canonicalise_factor_b(
-            group["B"], _factor_action_count(group["B"])
-        )
-        a_lists.append(np.asarray(group["A"], dtype=np.float64).tolist())
-        b_lists.append(b_canonical.tolist())
-        c_lists.append(np.asarray(group["C"], dtype=np.float64).flatten().tolist())
-        d_lists.append(np.asarray(group["D"], dtype=np.float64).flatten().tolist())
-
-    model_params = gnn_spec.get("model_parameters", {}) or {}
-    num_timesteps = int(
-        model_params.get("num_timesteps")
-        or gnn_spec.get("initialparameterization", {}).get("num_timesteps")
-        or 20
-    )
-    seed = int((options or {}).get("seed", 42))
-    action_precision = float((options or {}).get("action_precision", 4.0))
-
-    joint_size = 1
-    for factor_list in a_lists:
-        joint_size *= len(factor_list[0])
-
-    # nosec
-    code = f'''#!/usr/bin/env python3
-"""
-Kronecker-factorized JAX active inference for: {model_name}
-
-Generated from a GNN specification (roadmap MAJ-02). This script runs sparse
-mean-field active inference over per-factor matrices — the joint state space
-({joint_size} states) is reported in the results but never allocated. It
-requires ``jax`` and ``numpy`` and the GNN repository on the path.
-
-Results are written as ``simulation_results.json`` (schema
-``jax_kronecker_factorized_v1``) under ``GNN_OUTPUT_DIR`` (Step 12 sets it
-for the ``jax`` framework; defaults to a local output path otherwise).
-"""
-from __future__ import annotations
-
-import json
-import logging
-import os
-import sys
-from pathlib import Path
-
-# ---------------------------------------------------------------------------
-# Repository root resolution (prefer GNN_PROJECT_ROOT; else walk upwards)
-# ---------------------------------------------------------------------------
-_gnn_root = os.environ.get("GNN_PROJECT_ROOT")
-if _gnn_root:
-    _repo = Path(_gnn_root).resolve()
-    sys.path.insert(0, str(_repo / "src"))
-else:
-    _cur = Path(__file__).resolve().parent
-    _found = None
-    for _ in range(24):
-        if (_cur / "pyproject.toml").is_file() and (_cur / "src").is_dir():
-            _found = _cur
-            break
-        if _cur.parent == _cur:
-            break
-        _cur = _cur.parent
-    if _found is None:
-        print(
-            "ERROR: Cannot locate GNN repository root. Run via the pipeline "
-            "execute step, or set GNN_PROJECT_ROOT to the checkout root.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    sys.path.insert(0, str(_found / "src"))
-
-from gnn.execute.jax.kronecker_factorized import (  # noqa: E402
-    FactorizedPOMDP,
-    run_factorized_active_inference,
+from .jax_combined_generator import (
+    _generate_jax_combined_code,
 )
-
-logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
-logger = logging.getLogger(__name__)
-
-
-def main() -> int:
-    """Run sparse mean-field active inference for the embedded factor model."""
-    model = FactorizedPOMDP(
-        A={a_lists!r},
-        B={b_lists!r},
-        C={c_lists!r},
-        D={d_lists!r},
-        T={num_timesteps},
-        seed={seed},
-        action_precision={action_precision},
-    )
-    results = run_factorized_active_inference(model)
-
-    out_dir = Path(
-        os.environ.get(
-            "GNN_OUTPUT_DIR",
-            "output/jax_simulations/{safe_model_name}/simulation_data",
-        )
-    )
-    out_dir.mkdir(parents=True, exist_ok=True)
-    results_path = out_dir / "simulation_results.json"
-    results_path.write_text(json.dumps(results, indent=2, default=str))
-
-    logger.info("Simulation completed successfully")
-    logger.info("  framework:    %s", results.get("framework"))
-    logger.info("  model_kind:   %s", results.get("model_kind"))
-    logger.info("  num_factors:  %s", results.get("num_factors"))
-    logger.info("  joint_states: %s", results.get("model_parameters", {{}}).get("joint_state_space_size"))
-    logger.info("  timesteps:    %s", results.get("num_timesteps"))
-    logger.info("  results:      %s", results_path)
-    return 0
-
-
-if __name__ == "__main__":
-    sys.exit(main())
-'''
-    return code
-
-
-def _factor_action_count(value: Any) -> int:
-    """Per-factor action count from an action-major raw B tensor."""
-    raw = np.asarray(value, dtype=np.float64)
-    if raw.ndim == 2:
-        return 1
-    if raw.ndim == 3 and raw.shape[1] == raw.shape[2]:
-        return max(1, int(raw.shape[0]))
-    if raw.ndim == 3:
-        return max(1, int(raw.shape[-1]))
-    return 1
-
-
-# --- Internal code generation helpers ---
-
-
-def _parse_gnn_matrix_string(matrix_str: str) -> np.ndarray:
-    """Parse GNN matrix string format to numpy array."""
-    try:
-        # Remove comments and clean up
-        lines = matrix_str.split("\n")
-        cleaned_lines: list[Any] = []
-        for line in lines:
-            if "#" in line:
-                line = line.split("#")[0]
-            line = line.strip()
-            if line:
-                cleaned_lines.append(line)
-
-        # Reconstruct the matrix string
-        matrix_str = " ".join(cleaned_lines)
-
-        # Handle A matrix format: { (1.0, 0.0, 0.0), (0.0, 1.0, 0.0), (0.0, 0.0, 1.0) }
-        if matrix_str.startswith("{") and matrix_str.endswith("}"):
-            inner = matrix_str[1:-1].strip()
-
-            # Split by commas, but be careful with nested tuples
-            rows: list[Any] = []
-            current_row = ""
-            paren_count = 0
-
-            for char in inner:
-                if char == "(":
-                    paren_count += 1
-                elif char == ")":
-                    paren_count -= 1
-
-                if char == "," and paren_count == 0:
-                    # End of a row
-                    if current_row.strip():
-                        rows.append(current_row.strip())
-                    current_row = ""
-                else:
-                    current_row += char
-
-            # Add the last row
-            if current_row.strip():
-                rows.append(current_row.strip())
-
-            # Parse each row
-            matrix: list[Any] = []
-            for row in rows:
-                row = row.strip()
-                try:
-                    if row.startswith("(") and row.endswith(")"):
-                        # Parse tuple row
-                        inner_row = row[1:-1]
-                        row_values = [
-                            float(x.strip()) for x in inner_row.split(",") if x.strip()
-                        ]
-                        matrix.append(row_values)
-                    elif row.startswith("((") and row.endswith("))"):
-                        # Parse nested tuple row (for B matrix)
-                        inner_row = row[2:-2]
-                        # Split by '),('
-                        nested_tuples = inner_row.split("),(")
-                        nested_row_values: list[list[float]] = []
-                        for nested_tuple in nested_tuples:
-                            nested_tuple = nested_tuple.strip("()")
-                            tuple_values = [
-                                float(x.strip())
-                                for x in nested_tuple.split(",")
-                                if x.strip()
-                            ]
-                            nested_row_values.append(tuple_values)
-                        matrix.append(nested_row_values)
-                    else:
-                        # Try to parse as simple values
-                        row_values = [
-                            float(x.strip()) for x in row.split(",") if x.strip()
-                        ]
-                        if row_values:
-                            matrix.append(row_values)
-                except Exception as e:
-                    logger.warning(f"Failed to parse row '{row}': {e}")
-                    # Add a default row to maintain matrix structure
-                    if matrix:
-                        # Use the same length as previous rows
-                        matrix.append([0.0] * len(matrix[0]))
-                    else:
-                        matrix.append([1.0])
-
-            if not matrix:
-                return np.array([[1.0]])
-
-            # Ensure all rows have the same length
-            max_len = max(len(row) for row in matrix)
-            for _, row in enumerate(matrix):
-                while len(row) < max_len:
-                    row.append(0.0)
-
-            return np.array(matrix)
-
-        return np.array([[1.0]])  # Default recovery
-
-    except Exception as e:
-        logger.warning(f"Failed to parse matrix string: {e}")
-        return np.array([[1.0]])  # Default recovery
-
-
-def _extract_gnn_matrices(gnn_spec: Dict[str, Any]) -> Dict[str, Any]:
-    """Extract A, B, C, D matrices from GNN specification."""
-    matrices: dict[Any, Any] = {}
-
-    # --- Primary: Handle POMDP processor format ---
-    if "model_parameters" in gnn_spec:
-        logger.info("Extracting matrices from POMDP processor format")
-        init_candidate = gnn_spec.get("initialparameterization", {})
-        if isinstance(init_candidate, dict) and all(
-            key in init_candidate for key in ("A", "B", "C", "D")
-        ):
-            try:
-                gnn_spec = build_canonical_pomdp_spec(gnn_spec)
-                logger.info("Canonicalized POMDP matrices for JAX renderer")
-            except Exception as exc:
-                logger.warning(
-                    "Could not canonicalize POMDP matrices for JAX renderer: %s",
-                    exc,
-                )
-
-        model_params = gnn_spec["model_parameters"]
-        n_states = model_params.get("num_hidden_states", 3)
-        n_obs = model_params.get("num_obs", 3)
-        n_actions = model_params.get("num_actions", 3)
-
-        logger.info(
-            f"Extracted variable dimensions: {{n_states: {n_states}, n_obs: {n_obs}, n_actions: {n_actions}}}"
-        )
-
-        # Create default matrices based on dimensions
-        default_matrices: dict[str, Any] = {
-            "A": np.eye(n_obs, n_states),  # Identity-like likelihood matrix
-            "B": np.stack(
-                [np.eye(n_states) for _ in range(n_actions)], axis=2
-            ),  # Identity transitions for each action
-            "C": np.zeros(n_obs),  # Zero preferences
-            "D": np.ones(n_states) / n_states,  # Uniform prior
-        }
-        logger.info(
-            f"Created default matrices: A={default_matrices['A'].shape}, B={default_matrices['B'].shape}, C={default_matrices['C'].shape}, D={default_matrices['D'].shape}"
-        )
-
-        # Initialize matrices with defaults
-        matrices.update(default_matrices)
-
-        # Extract actual parameter values from initialparameterization
-        init_params = gnn_spec.get("initialparameterization", {})
-
-        # Override dimensions from B matrix if available (consistent with other renderers)
-        if "B" in init_params:
-            B_matrix = init_params["B"]
-            if isinstance(B_matrix, (list, np.ndarray)) and len(B_matrix) > 0:
-                B_array = np.asarray(B_matrix, dtype=float)
-                n_actions_from_b = B_array.shape[2] if B_array.ndim >= 3 else 1
-                n_actions = n_actions_from_b
-                logger.info(f"Corrected n_actions from B matrix: {n_actions}")
-
-            # Also check for explicit dimensions in model_params that might override
-            if "B" in gnn_spec.get("model_params", {}):
-                B_spec = gnn_spec["model_params"]["B"]
-                if "shape" in B_spec:
-                    shape_parts = B_spec["shape"].strip("()").split(",")
-                    if len(shape_parts) >= 3:
-                        try:
-                            n_actions_from_spec = int(shape_parts[2])
-                            if n_actions_from_spec > 1:
-                                n_actions = n_actions_from_spec
-                                logger.info(
-                                    f"Corrected n_actions from B matrix specification: {n_actions}"
-                                )
-                        except (ValueError, IndexError):
-                            logger.debug(
-                                "Could not extract n_actions from B matrix spec shape_parts"
-                            )
-        if init_params:
-            logger.info(
-                "Found initialparameterization, extracting actual matrix values"
-            )
-
-            # Extract A matrix
-            if "A" in init_params:
-                try:
-                    A_data = init_params["A"]
-                    # Handle both list and tuple (POMDP extractor returns tuples)
-                    if isinstance(A_data, (list, tuple)):
-                        A_matrix = np.array(A_data)
-                        if A_matrix.ndim == 2:
-                            matrices["A"] = A_matrix
-                            logger.info(
-                                f"Successfully extracted A matrix: shape {A_matrix.shape}"
-                            )
-                except Exception as e:
-                    logger.warning(f"Failed to extract A matrix: {e}")
-
-            # Extract B matrix
-            if "B" in init_params:
-                try:
-                    B_data = init_params["B"]
-                    # Handle both list and tuple (POMDP extractor returns tuples)
-                    if isinstance(B_data, (list, tuple)):
-                        B_matrix = np.array(B_data)
-                    elif isinstance(B_data, np.ndarray):
-                        B_matrix = B_data
-                    else:
-                        B_matrix = np.array([])
-                    if B_matrix.ndim == 2:
-                        matrices["B"] = B_matrix[:, :, np.newaxis]
-                        logger.info(
-                            f"Successfully extracted passive B matrix: shape {matrices['B'].shape}"
-                        )
-                    elif B_matrix.ndim == 3:
-                        matrices["B"] = B_matrix
-                        logger.info(
-                            f"Successfully extracted B matrix: shape {B_matrix.shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to extract B matrix: {e}")
-
-            # Extract C vector
-            if "C" in init_params:
-                try:
-                    C_data = init_params["C"]
-                    if isinstance(C_data, list):
-                        C_vector = np.array(C_data)
-                        if C_vector.ndim == 1:
-                            matrices["C"] = C_vector
-                            logger.info(
-                                f"Successfully extracted C vector: shape {C_vector.shape}"
-                            )
-                except Exception as e:
-                    logger.warning(f"Failed to extract C vector: {e}")
-
-            # Extract D vector
-            if "D" in init_params:
-                try:
-                    D_data = init_params["D"]
-                    if isinstance(D_data, list):
-                        D_vector = np.array(D_data)
-                        if D_vector.ndim == 1:
-                            matrices["D"] = D_vector
-                            logger.info(
-                                f"Successfully extracted D vector: shape {D_vector.shape}"
-                            )
-                except Exception as e:
-                    logger.warning(f"Failed to extract D vector: {e}")
-
-    # --- Recovery: Handle the JSON export format from GNN processing pipeline ---
-    elif "statespaceblock" in gnn_spec:
-        logger.info("Extracting matrices from GNN JSON export structure")
-
-        # Extract variable dimensions from statespaceblock
-        var_dims: dict[Any, Any] = {}
-        for var_data in gnn_spec.get("statespaceblock", []):
-            var_name = var_data.get("id", "")
-            dimensions_str = var_data.get("dimensions", "")
-            # Parse dimensions like "3,3,type=float" -> [3, 3]
-            if dimensions_str:
-                dims_parts = dimensions_str.split(",")
-                dims: list[Any] = []
-                for part in dims_parts:
-                    part = part.strip()
-                    if part.startswith("type="):
-                        break
-                    try:
-                        dims.append(int(part))
-                    except ValueError:
-                        logger.debug("Skipping non-integer dimension token: %s", part)
-                        continue
-                var_dims[var_name] = dims
-
-        # Create default matrices based on dimensions
-        default_matrices = {}
-        if "A" in var_dims:
-            dims = var_dims["A"]
-            if len(dims) >= 2:
-                default_matrices["A"] = np.eye(dims[0], dims[1])  # Identity matrix
-                logger.info(f"Created default A matrix with dimensions {dims}")
-
-        if "B" in var_dims:
-            dims = var_dims["B"]
-            if len(dims) >= 3:
-                # Create identity-like transition matrix
-                default_matrices["B"] = np.eye(dims[0], dims[1])[:, :, np.newaxis]
-                default_matrices["B"] = np.repeat(
-                    default_matrices["B"], dims[2], axis=2
-                )
-                logger.info(f"Created default B matrix with dimensions {dims}")
-
-        if "C" in var_dims:
-            dims = var_dims["C"]
-            if len(dims) >= 1:
-                default_matrices["C"] = np.zeros(dims[0])  # Zero preferences
-                logger.info(f"Created default C vector with dimensions {dims}")
-
-        if "D" in var_dims:
-            dims = var_dims["D"]
-            if len(dims) >= 1:
-                default_matrices["D"] = np.ones(dims[0]) / dims[0]  # Uniform prior
-                logger.info(f"Created default D vector with dimensions {dims}")
-
-        # Initialize matrices with defaults
-        matrices.update(default_matrices)
-
-        # Extract actual parameter values from InitialParameterization
-        initial_params = gnn_spec.get("raw_sections", {}).get(
-            "InitialParameterization", ""
-        )
-        if initial_params:
-            # Parse A matrix
-            a_match = re.search(r"A\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if a_match:
-                try:
-                    a_str = a_match.group(1).strip()
-                    parsed_matrix = _parse_gnn_matrix_string(f"{{{a_str}}}")
-                    if parsed_matrix.shape != (1, 1):
-                        matrices["A"] = parsed_matrix
-                        logger.info(
-                            f"Successfully parsed A matrix from InitialParameterization: shape {parsed_matrix.shape}"
-                        )
-                    elif "A" in default_matrices:
-                        improved_matrix = _create_improved_default_matrix(
-                            "A", default_matrices["A"], a_str
-                        )
-                        matrices["A"] = improved_matrix
-                        logger.info(
-                            f"Used improved default A matrix: shape {improved_matrix.shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse A matrix: {e}")
-
-            # Parse B matrix
-            b_match = re.search(r"B\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if b_match:
-                try:
-                    b_str = b_match.group(1).strip()
-                    parsed_matrix = _parse_gnn_matrix_string(f"{{{b_str}}}")
-                    if parsed_matrix.shape != (1, 1):
-                        matrices["B"] = parsed_matrix
-                        logger.info(
-                            f"Successfully parsed B matrix from InitialParameterization: shape {parsed_matrix.shape}"
-                        )
-                    elif "B" in default_matrices:
-                        improved_matrix = _create_improved_default_matrix(
-                            "B", default_matrices["B"], b_str
-                        )
-                        matrices["B"] = improved_matrix
-                        logger.info(
-                            f"Used improved default B matrix: shape {improved_matrix.shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse B matrix: {e}")
-
-            # Parse C vector
-            c_match = re.search(r"C\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if c_match:
-                try:
-                    c_str = c_match.group(1).strip()
-                    parsed_vector = _parse_gnn_matrix_string(f"{{{c_str}}}")
-                    if parsed_vector.shape != (1, 1):
-                        matrices["C"] = parsed_vector.flatten()
-                        logger.info(
-                            f"Successfully parsed C vector from InitialParameterization: shape {parsed_vector.flatten().shape}"
-                        )
-                    elif "C" in default_matrices:
-                        improved_vector = _create_improved_default_matrix(
-                            "C", default_matrices["C"], c_str
-                        )
-                        matrices["C"] = improved_vector
-                        logger.info(
-                            f"Used improved default C vector: shape {improved_vector.shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse C vector: {e}")
-
-            # Parse D vector
-            d_match = re.search(r"D\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if d_match:
-                try:
-                    d_str = d_match.group(1).strip()
-                    parsed_vector = _parse_gnn_matrix_string(f"{{{d_str}}}")
-                    if parsed_vector.shape != (1, 1):
-                        matrices["D"] = parsed_vector.flatten()
-                        logger.info(
-                            f"Successfully parsed D vector from InitialParameterization: shape {parsed_vector.flatten().shape}"
-                        )
-                    elif "D" in default_matrices:
-                        improved_vector = _create_improved_default_matrix(
-                            "D", default_matrices["D"], d_str
-                        )
-                        matrices["D"] = improved_vector
-                        logger.info(
-                            f"Used improved default D vector: shape {improved_vector.shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse D vector: {e}")
-
-    # Handle parsed GNN data structure (older format)
-    elif "variables" in gnn_spec:
-        logger.info("Extracting matrices from parsed GNN data structure")
-
-        # Extract variable dimensions from the variables list
-        var_dims = {}
-        for _, var_data in enumerate(gnn_spec.get("variables", [])):
-            var_name = var_data.get("id", "")  # Use 'id' instead of 'name'
-            dimensions_str = var_data.get(
-                "dimensions", ""
-            )  # This is a string like "3,3,type=float"
-
-            # Parse dimensions string like "3,3,type=float" -> [3, 3]
-            if var_name and dimensions_str:
-                dims = []
-                for part in dimensions_str.split(","):
-                    part = part.strip()
-                    if part.startswith("type="):
-                        break
-                    try:
-                        dims.append(int(part))
-                    except ValueError:
-                        logger.debug("Skipping non-integer dimension token: %s", part)
-                        continue
-
-                if dims:  # Only add if we successfully parsed dimensions
-                    var_dims[var_name] = dims
-                    logger.info(f"Found variable '{var_name}' with dimensions {dims}")
-                else:
-                    logger.warning(
-                        f"Could not parse dimensions for variable '{var_name}': '{dimensions_str}'"
-                    )
-
-        logger.info(f"Extracted variable dimensions: {var_dims}")
-
-        # Create default matrices based on dimensions
-        default_matrices = {}
-        if "A" in var_dims:
-            dims = var_dims["A"]
-            if len(dims) >= 2:
-                default_matrices["A"] = np.eye(dims[0], dims[1])  # Identity matrix
-                logger.info(
-                    f"Created default A matrix with dimensions {dims} -> shape {default_matrices['A'].shape}"
-                )
-
-        if "B" in var_dims:
-            dims = var_dims["B"]
-            if len(dims) >= 3:
-                # Create identity-like transition matrix
-                default_matrices["B"] = np.eye(dims[0], dims[1])[:, :, np.newaxis]
-                default_matrices["B"] = np.repeat(
-                    default_matrices["B"], dims[2], axis=2
-                )
-                logger.info(
-                    f"Created default B matrix with dimensions {dims} -> shape {default_matrices['B'].shape}"
-                )
-
-        if "C" in var_dims:
-            dims = var_dims["C"]
-            if len(dims) >= 1:
-                default_matrices["C"] = np.zeros(dims[0])  # Zero preferences
-                logger.info(
-                    f"Created default C vector with dimensions {dims} -> shape {default_matrices['C'].shape}"
-                )
-
-        if "D" in var_dims:
-            dims = var_dims["D"]
-            if len(dims) >= 1:
-                default_matrices["D"] = np.ones(dims[0]) / dims[0]  # Uniform prior
-                logger.info(
-                    f"Created default D vector with dimensions {dims} -> shape {default_matrices['D'].shape}"
-                )
-
-        # Initialize matrices with defaults (will be overwritten if parameter parsing succeeds)
-        matrices.update(default_matrices)
-        logger.info(
-            f"Initialized matrices with shapes: A={matrices.get('A', 'None')}, B={matrices.get('B', 'None')}, C={matrices.get('C', 'None')}, D={matrices.get('D', 'None')}"
-        )
-
-        # Extract actual parameter values from InitialParameterization if available
-        initial_params = gnn_spec.get("InitialParameterization", "")
-        if initial_params:
-            logger.info(
-                "Found InitialParameterization section, attempting to parse matrix values"
-            )
-
-            # Parse A matrix
-            a_match = re.search(r"A\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if a_match:
-                try:
-                    a_str = a_match.group(1).strip()
-                    parsed_matrix = _parse_gnn_matrix_string(f"{{{a_str}}}")
-                    if parsed_matrix.shape != (1, 1):
-                        matrices["A"] = parsed_matrix
-                        logger.info(
-                            f"Successfully parsed A matrix from InitialParameterization: shape {parsed_matrix.shape}"
-                        )
-                    elif "A" in default_matrices:
-                        # Use default matrix with correct dimensions
-                        matrices["A"] = default_matrices["A"]
-                        logger.info(
-                            f"Used default A matrix: shape {default_matrices['A'].shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse A matrix: {e}")
-
-            # Parse B matrix
-            b_match = re.search(r"B\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if b_match:
-                try:
-                    b_str = b_match.group(1).strip()
-                    parsed_matrix = _parse_gnn_matrix_string(f"{{{b_str}}}")
-                    if parsed_matrix.shape != (1, 1):
-                        matrices["B"] = parsed_matrix
-                        logger.info(
-                            f"Successfully parsed B matrix from InitialParameterization: shape {parsed_matrix.shape}"
-                        )
-                    elif "B" in default_matrices:
-                        matrices["B"] = default_matrices["B"]
-                        logger.info(
-                            f"Used default B matrix: shape {default_matrices['B'].shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse B matrix: {e}")
-
-            # Parse C vector
-            c_match = re.search(r"C\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if c_match:
-                try:
-                    c_str = c_match.group(1).strip()
-                    parsed_vector = _parse_gnn_matrix_string(f"{{{c_str}}}")
-                    if parsed_vector.shape != (1, 1):
-                        matrices["C"] = parsed_vector.flatten()
-                        logger.info(
-                            f"Successfully parsed C vector from InitialParameterization: shape {parsed_vector.flatten().shape}"
-                        )
-                    elif "C" in default_matrices:
-                        matrices["C"] = default_matrices["C"]
-                        logger.info(
-                            f"Used default C vector: shape {default_matrices['C'].shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse C vector: {e}")
-
-            # Parse D vector
-            d_match = re.search(r"D\s*=\s*\{([^}]+)\}", initial_params, re.DOTALL)
-            if d_match:
-                try:
-                    d_str = d_match.group(1).strip()
-                    parsed_vector = _parse_gnn_matrix_string(f"{{{d_str}}}")
-                    if parsed_vector.shape != (1, 1):
-                        matrices["D"] = parsed_vector.flatten()
-                        logger.info(
-                            f"Successfully parsed D vector from InitialParameterization: shape {parsed_vector.flatten().shape}"
-                        )
-                    elif "D" in default_matrices:
-                        matrices["D"] = default_matrices["D"]
-                        logger.info(
-                            f"Used default D vector: shape {default_matrices['D'].shape}"
-                        )
-                except Exception as e:
-                    logger.warning(f"Failed to parse D vector: {e}")
-
-        # Extract actual parameter values if available (older parameters section)
-        for param_data in gnn_spec.get("parameters", []):
-            param_name = param_data.get("name", "")
-            param_value = param_data.get("value")
-
-            if param_name in ["A", "B", "C", "D"] and param_value is not None:
-                try:
-                    if isinstance(param_value, str):
-                        # Parse GNN matrix string format
-                        parsed_matrix = _parse_gnn_matrix_string(param_value)
-
-                        # Enhanced dimension inference and validation
-                        if parsed_matrix.shape != (1, 1):
-                            # Parsing succeeded with meaningful dimensions
-                            matrices[param_name] = parsed_matrix
-                            logger.info(
-                                f"Successfully parsed {param_name} matrix from string: shape {parsed_matrix.shape}"
-                            )
-                        elif param_name in default_matrices:
-                            # Parsing failed but we have default dimensions - use improved matrix
-                            improved_matrix = _create_improved_default_matrix(
-                                param_name, default_matrices[param_name], param_value
-                            )
-                            matrices[param_name] = improved_matrix
-                            logger.info(
-                                f"Used improved default {param_name} matrix based on context: shape {improved_matrix.shape}"
-                            )
-                        else:
-                            # Parsing failed and no defaults - try dimension inference from context
-                            inferred_matrix = _infer_matrix_from_context(
-                                param_name, param_value, var_dims
-                            )
-                            matrices[param_name] = inferred_matrix
-                            logger.info(
-                                f"Inferred {param_name} matrix from context: shape {inferred_matrix.shape}"
-                            )
-
-                    elif isinstance(param_value, (list, tuple)):
-                        matrices[param_name] = np.array(param_value)
-                        logger.info(f"Converted {param_name} list/tuple to array")
-                    elif isinstance(param_value, set):
-                        # Convert set to list then to array
-                        matrices[param_name] = np.array(list(param_value))
-                        logger.info(f"Converted {param_name} set to array")
-                    else:
-                        matrices[param_name] = param_value
-                        logger.info(f"Used {param_name} parameter value directly")
-                except Exception as e:
-                    logger.warning(f"Failed to convert {param_name} parameter: {e}")
-                    # Create recovery matrix based on parameter name and expected dimensions
-                    fallback_matrix = _create_fallback_matrix(param_name, var_dims)
-                    matrices[param_name] = fallback_matrix
-                    logger.info(
-                        f"Created recovery {param_name} matrix: shape {fallback_matrix.shape}"
-                    )
-                    continue
-
-    else:
-        # Handle older raw text format
-        logger.info("Extracting matrices from raw text format")
-
-        # Extract InitialParameterization section
-        init_params = gnn_spec.get("InitialParameterization", "")
-        if not init_params:
-            logger.warning("No InitialParameterization found in GNN spec")
-            return matrices
-
-        # Parse A matrix (observation model)
-        a_match = re.search(r"A\s*=\s*\[(.*?)\]", init_params, re.DOTALL)
-        if a_match:
-            try:
-                a_str = a_match.group(1).strip()
-                a_matrix = _parse_matrix_string(a_str)
-                matrices["A"] = a_matrix
-            except Exception as e:
-                logger.error(f"Failed to parse A matrix: {e}")
-
-        # Parse B matrix (transition model)
-        b_match = re.search(r"B\s*=\s*\[(.*?)\]", init_params, re.DOTALL)
-        if b_match:
-            try:
-                b_str = b_match.group(1).strip()
-                b_matrix = _parse_matrix_string(b_str)
-                matrices["B"] = b_matrix
-            except Exception as e:
-                logger.error(f"Failed to parse B matrix: {e}")
-
-        # Parse C vector (preferences)
-        c_match = re.search(r"C\s*=\s*\[(.*?)\]", init_params, re.DOTALL)
-        if c_match:
-            try:
-                c_str = c_match.group(1).strip()
-                c_vector = _parse_vector_string(c_str)
-                matrices["C"] = c_vector
-            except Exception as e:
-                logger.error(f"Failed to parse C vector: {e}")
-
-        # Parse D vector (priors)
-        d_match = re.search(r"D\s*=\s*\[(.*?)\]", init_params, re.DOTALL)
-        if d_match:
-            try:
-                d_str = d_match.group(1).strip()
-                d_vector = _parse_vector_string(d_str)
-                matrices["D"] = d_vector
-            except Exception as e:
-                logger.error(f"Failed to parse D vector: {e}")
-
-    return matrices
-
-
-def _validated_jax_matrices(gnn_spec: Dict[str, Any]) -> Dict[str, np.ndarray]:
-    """Extract a coherent dense POMDP matrix set or fail closed.
-
-    Generated code must reflect the supplied model. Historically incomplete
-    inputs fell through to unrelated 2-state defaults and, on any exception,
-    a recovery script that did no inference. Validate the contract before
-    interpolation so a successful render always contains the requested model.
-    """
-    extracted = _extract_gnn_matrices(gnn_spec)
-    missing = [name for name in ("A", "B", "C", "D") if name not in extracted]
-    if missing:
-        raise ValueError(
-            "JAX rendering requires canonical A/B/C/D parameters; missing "
-            + ", ".join(missing)
-        )
-
-    matrices = {
-        name: np.asarray(extracted[name], dtype=np.float64)
-        for name in ("A", "B", "C", "D")
-    }
-    matrices["C"] = matrices["C"].reshape(-1)
-    matrices["D"] = matrices["D"].reshape(-1)
-    a_matrix = matrices["A"]
-    b_matrix = matrices["B"]
-    c_vector = matrices["C"]
-    d_vector = matrices["D"]
-
-    if a_matrix.ndim != 2:
-        raise ValueError(f"A must be 2-D [observation, state], got {a_matrix.shape}")
-    if b_matrix.ndim != 3:
-        raise ValueError(
-            f"B must be 3-D [next_state, previous_state, action], got {b_matrix.shape}"
-        )
-    num_observations, num_states = a_matrix.shape
-    if b_matrix.shape[:2] != (num_states, num_states):
-        raise ValueError(
-            "B state axes must match A state count: "
-            f"A={a_matrix.shape}, B={b_matrix.shape}"
-        )
-    if c_vector.shape != (num_observations,):
-        raise ValueError(
-            f"C length must match A observations: A={a_matrix.shape}, C={c_vector.shape}"
-        )
-    if d_vector.shape != (num_states,):
-        raise ValueError(
-            f"D length must match A states: A={a_matrix.shape}, D={d_vector.shape}"
-        )
-    if b_matrix.shape[2] <= 0:
-        raise ValueError("B must declare at least one action")
-    if any(not np.all(np.isfinite(value)) for value in matrices.values()):
-        raise ValueError("JAX matrices must contain only finite values")
-    return matrices
-
-
-def _jax_model_name(gnn_spec: Dict[str, Any], fallback: str) -> str:
-    """Return a safe Python-facing model label from canonical or earlier keys."""
-    value = (
-        gnn_spec.get("model_name")
-        or gnn_spec.get("ModelName")
-        or gnn_spec.get("name")
-        or fallback
-    )
-    return str(value).replace(" ", "_")
-
-
-def _create_improved_default_matrix(
-    param_name: str, default_matrix: np.ndarray, param_value: str
-) -> np.ndarray:
-    """Create an improved default matrix based on context clues from the failed parsing."""
-    # Try to extract numerical values from the failed parsing string
-    import re
-
-    # Look for numerical patterns in the string
-    numbers = re.findall(r"-?\d+\.?\d*", param_value)
-
-    if numbers and len(numbers) > 1:
-        # We found numbers, try to use them to improve the default matrix
-        try:
-            float_numbers = [float(n) for n in numbers]
-
-            if param_name == "A":
-                # For A matrix, try to create observation model with extracted values
-                shape = default_matrix.shape
-                new_matrix = np.zeros(shape)
-                for i, val in enumerate(float_numbers[: np.prod(shape)]):
-                    row = i // shape[1]
-                    col = i % shape[1]
-                    if row < shape[0] and col < shape[1]:
-                        new_matrix[row, col] = val
-
-                # Normalize rows to make it a proper probability matrix
-                row_sums = new_matrix.sum(axis=1, keepdims=True)
-                row_sums = np.where(row_sums > 0, row_sums, 1.0)
-                new_matrix = new_matrix / row_sums
-                return new_matrix
-
-            elif param_name == "B":
-                # For B matrix, create transition model with extracted values
-                shape = default_matrix.shape
-                new_matrix = np.zeros(shape)
-                for i, val in enumerate(float_numbers[: np.prod(shape)]):
-                    # Map linear index to 3D coordinates
-                    idx_2d = i % (shape[0] * shape[1])
-                    action = i // (shape[0] * shape[1])
-                    row = idx_2d // shape[1]
-                    col = idx_2d % shape[1]
-                    if action < shape[2] and row < shape[0] and col < shape[1]:
-                        new_matrix[row, col, action] = val
-
-                # Normalize each action matrix
-                for a in range(shape[2]):
-                    action_matrix = new_matrix[:, :, a]
-                    row_sums = action_matrix.sum(axis=1, keepdims=True)
-                    row_sums = np.where(row_sums > 0, row_sums, 1.0)
-                    new_matrix[:, :, a] = action_matrix / row_sums
-                return new_matrix
-
-            elif param_name in ["C", "D", "E"]:
-                # For vectors, use extracted values directly
-                shape = default_matrix.shape
-                new_vector = np.zeros(shape)
-                for i, val in enumerate(float_numbers[: shape[0]]):
-                    new_vector[i] = val
-
-                # Normalize if it's a probability vector (D, E)
-                if param_name in ["D", "E"]:
-                    vector_sum = new_vector.sum()
-                    if vector_sum > 0:
-                        new_vector = new_vector / vector_sum
-                    else:
-                        new_vector = np.ones(shape) / shape[0]
-
-                return new_vector
-        except (ValueError, TypeError, IndexError):
-            logger.debug(
-                "Could not improve matrix from parameterization for %s, using default",
-                param_name,
-            )
-
-    # If we can't improve it, return the default
-    return default_matrix
-
-
-def _infer_matrix_from_context(
-    param_name: str, param_value: str, var_dims: dict
-) -> np.ndarray:
-    """Infer matrix dimensions and create appropriate matrix when no defaults are available."""
-
-    # Try to infer dimensions from variable information
-    if param_name in var_dims:
-        dims = var_dims[param_name]
-    else:
-        # Use standard POMDP defaults
-        dims = {
-            "A": [2, 2],  # 2 observations x 2 states
-            "B": [2, 2, 2],  # 2 states x 2 states x 2 actions
-            "C": [2],  # 2 observations
-            "D": [2],  # 2 states
-            "E": [2],  # 2 actions
-        }.get(param_name, [2])
-
-    # Create appropriate matrix based on parameter type
-    if param_name == "A":
-        # Observation model - create informative but not deterministic
-        shape = tuple(dims) if len(dims) >= 2 else (2, 2)
-        matrix = np.eye(min(shape)) + 0.1 * np.random.rand(*shape)
-        # Normalize rows
-        matrix = matrix / matrix.sum(axis=1, keepdims=True)
-        return cast(np.ndarray, matrix)
-
-    elif param_name == "B":
-        # Transition model - create identity-like transitions with some noise
-        shape = tuple(dims) if len(dims) >= 3 else (2, 2, 2)
-        matrix = np.zeros(shape)
-        for a in range(shape[2]):
-            action_matrix = np.eye(shape[0], shape[1]) + 0.1 * np.random.rand(
-                shape[0], shape[1]
-            )
-            matrix[:, :, a] = action_matrix / action_matrix.sum(axis=1, keepdims=True)
-        return cast(np.ndarray, matrix)
-
-    elif param_name == "C":
-        # Preferences - slight preference for later observations
-        vector_len = int(dims[0]) if dims else 2
-        vector = np.linspace(0.1, 1.0, vector_len)
-        return cast(np.ndarray, vector)
-
-    elif param_name == "D":
-        # Prior - uniform
-        vector_len = int(dims[0]) if dims else 2
-        return cast(np.ndarray, np.ones(vector_len) / vector_len)
-
-    elif param_name == "E":
-        # Action prior - uniform
-        vector_len = int(dims[0]) if dims else 2
-        return cast(np.ndarray, np.ones(vector_len) / vector_len)
-
-    else:
-        # Generic recovery
-        vector_len = int(dims[0]) if dims else 2
-        return cast(np.ndarray, np.ones(vector_len) / vector_len)
-
-
-def _create_fallback_matrix(param_name: str, var_dims: dict) -> np.ndarray:
-    """Create a recovery matrix when all else fails."""
-    return _infer_matrix_from_context(param_name, "", var_dims)
-
-
-def _parse_matrix_string(matrix_str: str) -> np.ndarray:
-    """Parse matrix string to numpy array."""
-    # Remove extra whitespace and newlines
-    matrix_str = re.sub(r"\s+", " ", matrix_str.strip())
-
-    # Split by rows and parse each row
-    rows: list[Any] = []
-    for row_str in matrix_str.split(";"):
-        row_str = row_str.strip()
-        if row_str:
-            # Parse row as list of floats
-            row_values: list[Any] = []
-            for val_str in row_str.split(","):
-                val_str = val_str.strip()
-                if val_str:
-                    try:
-                        row_values.append(float(val_str))
-                    except ValueError:
-                        logger.warning(f"Could not parse value: {val_str}")
-                        row_values.append(0.0)
-            rows.append(row_values)
-
-    if not rows:
-        return np.array([[1.0]])
-
-    # Ensure all rows have same length
-    max_len = max(len(row) for row in rows)
-    for _, row in enumerate(rows):
-        while len(row) < max_len:
-            row.append(0.0)
-
-    return np.array(rows)
-
-
-def _parse_vector_string(vector_str: str) -> np.ndarray:
-    """Parse vector string to numpy array."""
-    # Remove extra whitespace
-    vector_str = re.sub(r"\s+", " ", vector_str.strip())
-
-    # Parse as list of floats
-    values: list[Any] = []
-    for val_str in vector_str.split(","):
-        val_str = val_str.strip()
-        if val_str:
-            try:
-                values.append(float(val_str))
-            except ValueError:
-                logger.warning(f"Could not parse value: {val_str}")
-                values.append(0.0)
-
-    if not values:
-        return np.array([1.0])
-
-    return np.array(values)
-
-
-def _json_dumps(value: str) -> str:
-    """Return *value* as a JSON string literal (module-level import site)."""
-    import json
-
-    return json.dumps(value)
-
-
-EFE_CONVENTION_JAX = (
-    "jax renderer compute_expected_free_energy heuristic:"
-    " EFE = obs_entropy - pragmatic_value + 0.1 * kl_divergence, where"
-    " obs_entropy is the entropy of predicted observations,"
-    " pragmatic_value = sum_o q(o) C[o] (a LINEAR payoff sum, not a KL"
-    " against C), and kl_divergence = KL(next_belief || D) weighted by a"
-    " fixed 0.1 coefficient. This is NOT the risk+ambiguity EFE"
-    " decomposition (risk is the KL against C) and is not comparable to"
-    " the pymdp neg_efe or the Lean"
-    " expectedFreeEnergy_eq_risk_add_ambiguity without convention"
-    " mapping (bridge finding O1)."
+from .jax_factorized_generator import (
+    _FACTOR_MATRIX_RE,
+    _canonicalise_factor_b,
+    _factor_action_count,
+    _factor_matrix_groups,
+    _generate_jax_factorized_code,
+    _is_factorized_spec,
 )
-EFE_CONVENTION_JAX_JSON = EFE_CONVENTION_JAX.replace("\n", " ")
-EFE_CONVENTION_JAX_JSON_LITERAL = _json_dumps(EFE_CONVENTION_JAX_JSON)
-
-
-def _generate_jax_model_code(
-    gnn_spec: Dict[str, Any], options: Optional[Dict[str, Any]]
-) -> str:
-    """Generate general JAX model code from GNN specification.
-
-    This generates standalone JAX code that only requires jax and jax.numpy,
-    without external dependencies like Flax or Optax.
-    """
-
-    try:
-        model_name = _jax_model_name(gnn_spec, "GNNModel")
-        matrices = _validated_jax_matrices(gnn_spec)
-
-        # Pre-compute dimensions to avoid f-string issues
-        A_matrix = matrices["A"]
-        B_matrix = matrices["B"]
-        C_vector = matrices["C"]
-        D_vector = matrices["D"]
-
-        num_states = A_matrix.shape[1]
-        num_observations = A_matrix.shape[0]
-        num_actions = B_matrix.shape[2]
-
-        # Extract num_timesteps from model_parameters.
-        model_params = gnn_spec.get("model_parameters", {})
-        init_params = gnn_spec.get("initialparameterization", {})
-        num_timesteps = model_params.get(
-            "num_timesteps", init_params.get("num_timesteps", 20)
-        )
-
-        # Convert matrices to lists for f-string insertion
-        A_list = A_matrix.tolist()
-        B_list = B_matrix.tolist()
-        C_list = C_vector.tolist()
-        D_list = D_vector.tolist()
-
-        # nosec
-        code = f'''"""
-JAX Model Generated from GNN Specification: {model_name}
-
-This model implements the GNN specification using pure JAX for high-performance computation.
-Generated automatically by the GNN Processing Pipeline.
-
-This is a standalone JAX implementation that only requires jax and numpy.
-No external dependencies like Flax or Optax are required.
-
-@Web: https://github.com/google/jax
-"""
-
-import jax
-import jax.numpy as jnp
-from jax import jit, vmap
-from functools import partial
-from typing import Dict, Any, Tuple
-import logging
-import numpy as np
-
-logger = logging.getLogger(__name__)
-
-
-# Model configuration
-NUM_STATES = {num_states}
-NUM_OBSERVATIONS = {num_observations}
-NUM_ACTIONS = {num_actions}
-
-
-def create_params() -> Dict[str, jnp.ndarray]:
-    """Create model parameters from GNN specification."""
-    return {{
-        'A_matrix': jnp.array({A_list}),  # Observation model P(o|s)
-        'B_matrix': jnp.array({B_list}),  # Transition model P(s'|s,a)
-        'C_vector': jnp.array({C_list}),  # Preferences over observations
-        'D_vector': jnp.array({D_list}),  # Prior over initial states
-    }}
-
-
-@jit
-def belief_update(params: Dict[str, jnp.ndarray], belief: jnp.ndarray, 
-                  observation: jnp.ndarray) -> jnp.ndarray:
-    """
-    Bayesian belief update given an observation.
-    
-    P(s|o) ∝ P(o|s) * P(s) using the A matrix (likelihood)
-    
-    Args:
-        params: Model parameters dictionary
-        belief: Current belief state [num_states]
-        observation: Observation vector [num_observations]
-        
-    Returns:
-        Updated belief state [num_states]
-    """
-    A_matrix = params['A_matrix']
-    
-    # Compute likelihood P(o|s) for each state
-    # For each state s, compute sum over observations: A[o,s] * observation[o]
-    likelihood = jnp.dot(observation, A_matrix)  # [num_states]
-    
-    # Bayesian update: P(s|o) ∝ P(o|s) * P(s)
-    updated_belief = belief * likelihood
-    
-    # Normalize with numerical stability
-    normalizer = jnp.sum(updated_belief) + 1e-8
-    updated_belief = updated_belief / normalizer
-    
-    return updated_belief
-
-
-@jit
-def compute_expected_free_energy(params: Dict[str, jnp.ndarray], belief: jnp.ndarray, 
-                                  action: int) -> float:
-    """
-    Compute expected free energy (EFE) for a given action.
-    
-    EFE = -E[log P(o)] + KL[Q(s')||P(s')]
-        = Entropy of predicted observations + KL divergence
-    
-    Args:
-        params: Model parameters dictionary
-        belief: Current belief state [num_states]
-        action: Action index
-        
-    Returns:
-        Expected free energy value
-    """
-    A_matrix = params['A_matrix']
-    B_matrix = params['B_matrix']
-    C_vector = params['C_vector']
-    D_vector = params['D_vector']
-    
-    # Predict next state distribution using B matrix (transitions)
-    # B_matrix[:, :, action] is [num_states, num_states]
-    next_belief = jnp.dot(B_matrix[:, :, action], belief)
-    next_belief = next_belief / (jnp.sum(next_belief) + 1e-8)  # Normalize
-    
-    # Predict observation distribution using A matrix
-    predicted_obs = jnp.dot(A_matrix, next_belief)
-    predicted_obs = predicted_obs / (jnp.sum(predicted_obs) + 1e-8)  # Normalize
-    
-    # Compute epistemic value (expected information gain)
-    # This is the negative entropy of predicted observations
-    obs_entropy = -jnp.sum(predicted_obs * jnp.log(predicted_obs + 1e-8))
-    
-    # Compute pragmatic value (expected utility based on preferences)
-    # Higher values in C_vector mean more preferred observations
-    pragmatic_value = jnp.dot(predicted_obs, C_vector)
-    
-    # Compute KL divergence between predicted state and prior
-    kl_divergence = jnp.sum(
-        jnp.where(next_belief > 1e-8,
-                  next_belief * jnp.log((next_belief + 1e-8) / (D_vector + 1e-8)),
-                  0.0)
-    )
-    
-    # Expected free energy = uncertainty - utility + complexity
-    efe = obs_entropy - pragmatic_value + 0.1 * kl_divergence
-    
-    return efe
-
-
-@jit
-def choose_action(params: Dict[str, jnp.ndarray], belief: jnp.ndarray) -> Tuple[int, jnp.ndarray]:
-    """
-    Choose action with minimum expected free energy.
-    
-    Args:
-        params: Model parameters dictionary
-        belief: Current belief state [num_states]
-        
-    Returns:
-        Tuple of (chosen_action_index, efe_values_for_all_actions)
-    """
-    # Compute EFE for all actions
-    efe_values = jnp.array([
-        compute_expected_free_energy(params, belief, a) 
-        for a in range(NUM_ACTIONS)
-    ])
-    
-    # Choose action with minimum EFE
-    chosen_action = jnp.argmin(efe_values)
-    
-    return chosen_action, efe_values
-
-# Batched operations for parallel execution across multiple agents or parallel rollouts
-# JAX's vmap allows automatic vectorization of functions
-batched_belief_update = jax.vmap(belief_update, in_axes=(None, 0, 0))
-batched_compute_expected_free_energy = jax.vmap(compute_expected_free_energy, in_axes=(None, 0, None))
-
-# Optional pmap for multiple devices if available
-import os
-try:
-    if len(jax.devices()) > 1 and os.environ.get("JAX_ENABLE_PMAP", "1") == "1":
-        # Multi-device data parallelism
-        parallel_belief_update = jax.pmap(belief_update, in_axes=(None, 0, 0))
-except (RuntimeError, ValueError):
-    logger.debug("pmap unavailable (single-device setup or runtime error)")
-
-
-@jit
-def state_transition(params: Dict[str, jnp.ndarray], belief: jnp.ndarray, 
-                     action: int) -> jnp.ndarray:
-    """
-    Predict next state distribution given current belief and action.
-    
-    Args:
-        params: Model parameters dictionary
-        belief: Current belief state [num_states]
-        action: Action index
-        
-    Returns:
-        Predicted next state distribution [num_states]
-    """
-    B_matrix = params['B_matrix']
-    
-    # Apply transition model
-    next_belief = jnp.dot(B_matrix[:, :, action], belief)
-    
-    # Normalize
-    next_belief = next_belief / (jnp.sum(next_belief) + 1e-8)
-    
-    return next_belief
-
-
-def simulate_step(params: Dict[str, jnp.ndarray], belief: jnp.ndarray, 
-                  observation: jnp.ndarray) -> Dict[str, Any]:
-    """
-    Perform one step of Active Inference simulation.
-    
-    This includes:
-    1. Belief update given observation
-    2. Action choice based on expected free energy
-    3. State prediction for chosen action
-    
-    Args:
-        params: Model parameters dictionary
-        belief: Current belief state [num_states]
-        observation: Observation vector [num_observations]
-        
-    Returns:
-        Dictionary with simulation results
-    """
-    # 1. Update belief based on observation
-    updated_belief = belief_update(params, belief, observation)
-    
-    # 2. Choose action
-    chosen_action, efe_values = choose_action(params, updated_belief)
-    
-    # 3. Predict next state
-    predicted_next_state = state_transition(params, updated_belief, chosen_action)
-    
-    return {{
-        'belief': updated_belief,
-        'action': chosen_action,
-        'expected_free_energy': efe_values[chosen_action],
-        'all_efe_values': efe_values,
-        'predicted_next_state': predicted_next_state,
-    }}
-
-
-def run_simulation(params: Dict[str, jnp.ndarray], num_steps: int, 
-                   initial_belief: jnp.ndarray = None, seed: int = 42) -> Dict[str, Any]:
-    """
-    Run a full Active Inference simulation interacting with a true
-    stochastic POMDP generative environment.
-    
-    Args:
-        params: Model parameters dictionary
-        num_steps: Number of simulation timesteps
-        initial_belief: Optional initial belief (default: uniform)
-        seed: Random seed for stochastic environment generation
-        
-    Returns:
-        Dictionary with simulation trajectory
-    """
-    key = jax.random.PRNGKey(seed)
-    
-    # Initialize belief
-    if initial_belief is None:
-        belief = params['D_vector'].copy()
-    else:
-        belief = initial_belief
-        
-    # Initialize true state from Prior
-    key, subkey = jax.random.split(key)
-    true_state_idx = jax.random.categorical(subkey, jnp.log(params['D_vector'] + 1e-8))
-    
-    # Storage for trajectory
-    beliefs = []
-    actions = []
-    efes = []
-    observations_log = []
-    
-    # Run simulation
-    for t in range(num_steps):
-        # 1. Environment generates observation
-        key, subkey = jax.random.split(key)
-        obs_probs = params['A_matrix'][:, true_state_idx]
-        obs_idx = jax.random.categorical(subkey, jnp.log(obs_probs + 1e-8))
-        
-        # Create one-hot observation for the agent
-        obs_one_hot = jnp.zeros(params['A_matrix'].shape[0])
-        obs_one_hot = obs_one_hot.at[obs_idx].set(1.0)
-        observations_log.append(obs_idx)
-        
-        # 2. Agent perceives and acts
-        result = simulate_step(params, belief, obs_one_hot)
-        
-        beliefs.append(result['belief'])
-        actions.append(result['action'])
-        efes.append(result['all_efe_values'])
-        action_idx = int(result['action'])
-        
-        # 3. Environment transitions true state
-        key, subkey = jax.random.split(key)
-        next_state_probs = params['B_matrix'][:, true_state_idx, action_idx]
-        true_state_idx = jax.random.categorical(subkey, jnp.log(next_state_probs + 1e-8))
-        
-        # Update belief for next step (using predicted next state)
-        belief = result['predicted_next_state']
-    
-    return {{
-        'beliefs': jnp.stack(beliefs),
-        'actions': jnp.array(actions),
-        'expected_free_energies': jnp.array(efes),
-        'observations': jnp.array(observations_log),
-        'final_belief': belief,
-    }}
-
-
-def get_model_summary() -> str:
-    """Get a summary of the model architecture."""
-    return f"""
-{model_name} Model Summary (Pure JAX Implementation):
-- Number of states: {{NUM_STATES}}
-- Number of observations: {{NUM_OBSERVATIONS}}
-- Number of actions: {{NUM_ACTIONS}}
-- Parameters: A matrix, B matrix, C vector, D vector
-
-Key Functions:
-- create_params(): Create model parameters
-- belief_update(): Bayesian belief update
-- compute_expected_free_energy(): Compute EFE for action
-- choose_action(): Choose action with minimum EFE
-- simulate_step(): One step of Active Inference
-- run_simulation(): Full simulation over observations
-"""
-
-
-def save_simulation_results(trajectory: Dict[str, Any], params: Dict[str, jnp.ndarray],
-                            model_name: str, output_dir: str = ".") -> str:
-    """
-    Save simulation results to structured JSON for downstream analysis.
-
-    Args:
-        trajectory: Simulation trajectory dictionary
-        params: Model parameters
-        model_name: Name of the model
-        output_dir: Output directory for JSON file
-
-    Returns:
-        Path to the saved JSON file
-    """
-    import json
-    import os
-    from datetime import datetime
-
-    # Create output directory if needed
-    os.makedirs(output_dir, exist_ok=True)
-
-    # Build structured results matching PyMDP format for cross-framework analysis
-    results = {{
-        "success": True,
-        "framework": "jax",
-        "model_name": model_name,
-        "num_timesteps": int(len(trajectory['actions'])),
-        "timestamp": datetime.now().isoformat(),
-        "simulation_trace": {{
-            "observations": trajectory.get('observations', jnp.array([])).tolist(),
-            "beliefs": [b.tolist() for b in trajectory.get('beliefs', [])],
-            "actions": trajectory.get('actions', jnp.array([])).tolist(),
-            "efe_history": trajectory.get('expected_free_energies', jnp.array([])).tolist(),
-            "belief_confidence": [float(max(b)) if len(b) > 0 else 0.0 for b in trajectory.get('beliefs', [])],
-        }},
-        "observations": trajectory.get('observations', jnp.array([])).tolist(),
-        "beliefs": [b.tolist() for b in trajectory['beliefs']],
-        "actions": trajectory['actions'].tolist(),
-        "final_belief": trajectory['final_belief'].tolist(),
-        "model_parameters": {{
-            "A_shape": list(params['A_matrix'].shape),
-            "B_shape": list(params['B_matrix'].shape),
-            "C_shape": list(params['C_vector'].shape),
-            "D_shape": list(params['D_vector'].shape),
-            "num_states": NUM_STATES,
-            "num_observations": NUM_OBSERVATIONS,
-            "num_actions": NUM_ACTIONS
-        }},
-        "metrics": {{
-            "expected_free_energy": trajectory['expected_free_energies'].tolist(),
-            "expected_free_energy_convention": {EFE_CONVENTION_JAX_JSON_LITERAL},
-            "average_efe": float(jnp.mean(trajectory['expected_free_energies'])),
-            "belief_confidence": [float(max(b)) for b in trajectory['beliefs']],
-        }},
-        "validation": {{
-            "all_beliefs_valid": all(abs(sum(b) - 1.0) < 0.01 for b in trajectory['beliefs']),
-            "beliefs_sum_to_one": True,
-            "actions_in_range": all(0 <= a < NUM_ACTIONS for a in trajectory['actions'])
-        }}
-    }}
-
-    # Save to JSON
-    output_file = os.path.join(output_dir, "simulation_results.json")
-    with open(output_file, 'w') as f:
-        json.dump(results, f, indent=2)
-
-    return output_file
-
-
-if __name__ == "__main__":
-    import os
-
-    print("=" * 60)
-    print("JAX Active Inference Model: {model_name}")
-    print("=" * 60)
-
-    # Create parameters
-    params = create_params()
-    print("\\n✅ Model parameters created")
-    print(f"   A matrix shape: {{params['A_matrix'].shape}}")
-    print(f"   B matrix shape: {{params['B_matrix'].shape}}")
-    print(f"   C vector shape: {{params['C_vector'].shape}}")
-    print(f"   D vector shape: {{params['D_vector'].shape}}")
-
-    # Print model summary
-    print(get_model_summary())
-
-    # Test with uniform initial belief
-    print("\\n🧪 Running test simulation...")
-    initial_belief = params['D_vector']
-
-    # Create a test observation (one-hot for first observation)
-    test_obs = jnp.zeros(NUM_OBSERVATIONS)
-    test_obs = test_obs.at[0].set(1.0)
-
-    # Run one simulation step
-    result = simulate_step(params, initial_belief, test_obs)
-
-    print(f"\\n📊 Simulation Results:")
-    print(f"   Initial belief: {{initial_belief}}")
-    print(f"   Updated belief: {{result['belief']}}")
-    print(f"   Selected action: {{result['action']}}")
-    print(f"   Expected free energy: {{result['expected_free_energy']:.4f}}")
-    print(f"   EFE for all actions: {{result['all_efe_values']}}")
-
-    # Run multi-step simulation
-    print("\\n🔄 Running {num_timesteps}-step simulation...")
-    trajectory = run_simulation(params, num_steps={num_timesteps})
-
-    print(f"   Actions taken: {{trajectory['actions']}}")
-    print(f"   Final belief: {{trajectory['final_belief']}}")
-    print(f"   Average EFE: {{jnp.mean(trajectory['expected_free_energies']):.4f}}")
-
-    # Save structured results for analysis step
-    output_dir = os.environ.get('GNN_OUTPUT_DIR', 'jax_outputs')
-    results_file = save_simulation_results(trajectory, params, "{model_name}", output_dir)
-    print(f"\\n💾 Saved simulation results to: {{results_file}}")
-
-    print("\\n✅ JAX Active Inference model test successful!")
-'''
-
-        return code
-
-    except Exception as e:
-        raise ValueError(f"JAX model generation failed: {e}") from e
-
-
-def _generate_jax_pomdp_code(
-    gnn_spec: Dict[str, Any], options: Optional[Dict[str, Any]]
-) -> str:
-    """Generate JAX POMDP solver code from GNN specification."""
-
-    try:
-        model_name = _jax_model_name(gnn_spec, "POMDPModel")
-        matrices = _validated_jax_matrices(gnn_spec)
-        A_matrix = matrices["A"]
-        B_matrix = matrices["B"]
-        C_vector = matrices["C"]
-        D_vector = matrices["D"]
-
-        logger.info(f"A matrix shape: {A_matrix.shape}")
-        logger.info(f"B matrix shape: {B_matrix.shape}")
-        logger.info(f"C vector shape: {C_vector.shape}")
-        logger.info(f"D vector shape: {D_vector.shape}")
-
-        num_states = A_matrix.shape[1]
-        num_observations = A_matrix.shape[0]
-        num_actions = B_matrix.shape[2]
-
-        logger.info(
-            f"Final dimensions: states={num_states}, observations={num_observations}, actions={num_actions}"
-        )
-
-        code = f'''"""
-JAX POMDP Solver Generated from GNN Specification: {model_name}
-
-This implements a complete POMDP solver using JAX optimizations including JIT, vmap, and pmap.
-Based on the GNN specification with belief updates, value iteration, and alpha vector backup.
-
-@Web: https://pfjax.readthedocs.io
-@Web: https://arxiv.org/abs/1304.1118
-@Web: https://www.cs.cmu.edu/~ggordon/jpineau-ggordon-thrun.ijcai03.pdf
-"""
-
-import jax
-import jax.numpy as jnp
-from functools import partial
-from jax import jit, vmap, pmap
-from typing import Dict, Any, Optional, Tuple, List
-import numpy as np
-
-class POMDPModels:
-    """Container for POMDP model parameters."""
-    
-    def __init__(self, A: jnp.ndarray, B: jnp.ndarray, C: jnp.ndarray, D: jnp.ndarray):
-        self.A = A  # Observation model P(o|s)
-        self.B = B  # Transition model P(s'|s,u)
-        self.C = C  # Preferences over observations
-        self.D = D  # Prior over initial states
-        self.discount_factor = 0.95
-
-class JAXPOMDPSolver:
-    """
-    High-performance POMDP solver using JAX optimizations.
-    
-    Implements belief updates, value iteration, and alpha vector backup with JIT compilation.
-    """
-    
-    def __init__(self, models: POMDPModels):
-        self.models = models
-        self.num_states = models.A.shape[1]
-        self.num_observations = models.A.shape[0]
-        self.num_actions = models.B.shape[2]
-        self.discount_factor = models.discount_factor  # Add missing discount factor
-        
-        # JIT-compiled functions for performance
-        self.belief_update_jitted = jit(self.belief_update)
-        self.alpha_vector_backup_jitted = jit(self.alpha_vector_backup)
-    
-    @partial(jit, static_argnums=(0,))
-    def belief_update(self, belief: jnp.ndarray, action: int, observation: int) -> jnp.ndarray:
-        """
-        Bayesian belief update with numerical stability.
-        
-        Args:
-            belief: Current belief state
-            action: Action taken
-            observation: Observation received
-            
-        Returns:
-            Updated belief state
-        """
-        # Prediction step
-        predicted_belief = jnp.dot(self.models.B[:, :, action], belief)
-        
-        # Update step
-        updated_belief = predicted_belief * self.models.A[observation, :]
-        
-        # Normalization with numerical stability
-        normalizer = jnp.sum(updated_belief)
-        normalized_belief = jnp.where(
-            normalizer > 1e-10,
-            updated_belief / normalizer,
-            jnp.ones_like(updated_belief) / self.num_states
-        )
-        
-        return normalized_belief
-    
-    @partial(jit, static_argnums=(0,))
-    def alpha_vector_backup(self, belief: jnp.ndarray, action: int,
-                           alpha_vectors: jnp.ndarray) -> jnp.ndarray:
-        """
-        Optimized alpha vector backup with vectorization.
-        
-        Args:
-            belief: Current belief state
-            action: Action to evaluate
-            alpha_vectors: Current alpha vectors
-            
-        Returns:
-            New alpha vector for this action
-        """
-        # Convert observation preferences to state-conditioned immediate value.
-        base_alpha = jnp.dot(self.models.A.T, self.models.C)
-        
-        # Pre-calculate observation probabilities
-        next_belief_pred = jnp.dot(self.models.B[:, :, action], belief)
-        obs_probs = jnp.dot(self.models.A, next_belief_pred)
-        
-        def compute_obs_contribution(obs_idx):
-            next_belief = self.belief_update(belief, action, obs_idx)
-            values = jnp.dot(alpha_vectors, next_belief)
-            best_alpha = alpha_vectors[jnp.argmax(values)]
-            
-            # Only add contribution if observation probability is significant
-            return jnp.where(obs_probs[obs_idx] > 1e-10, 
-                           self.discount_factor * obs_probs[obs_idx] * best_alpha,
-                           jnp.zeros_like(best_alpha))
-                           
-        # Vectorized map-reduce over observations
-        contributions = vmap(compute_obs_contribution)(jnp.arange(self.num_observations))
-        alpha = base_alpha + jnp.sum(contributions, axis=0)
-        
-        return alpha
-    
-    def compute_observation_probability(self, belief: jnp.ndarray, action: int) -> jnp.ndarray:
-        """Compute probability of observations given belief and action."""
-        next_belief = jnp.dot(self.models.B[:, :, action], belief)
-        return jnp.dot(self.models.A, next_belief)
-
-def create_pomdp_solver() -> JAXPOMDPSolver:
-    """Create and return a POMDP solver with the specified model parameters."""
-    
-    # Model parameters from GNN specification
-    A = jnp.array({A_matrix.tolist()})  # Observation model
-    B = jnp.array({B_matrix.tolist()})  # Transition model  
-    C = jnp.array({C_vector.tolist()})  # Preferences
-    D = jnp.array({D_vector.tolist()})  # Prior
-    
-    models = POMDPModels(A=A, B=B, C=C, D=D)
-    return JAXPOMDPSolver(models)
-
-def solve_pomdp(solver: JAXPOMDPSolver, initial_belief: jnp.ndarray, 
-                horizon: int = 10) -> Dict[str, jnp.ndarray]:
-    """
-    Solve POMDP using value iteration with alpha vectors.
-    
-    Args:
-        solver: POMDP solver instance
-        initial_belief: Initial belief state
-        horizon: Planning horizon
-        
-    Returns:
-        Dictionary containing solution components
-    """
-    # Initialize alpha vectors
-    alpha_vectors = jnp.zeros((solver.num_actions, solver.num_states))
-    
-    # Value iteration
-    for t in range(horizon):
-        new_alpha_vectors = []
-        for action in range(solver.num_actions):
-            alpha = solver.alpha_vector_backup_jitted(initial_belief, action, alpha_vectors)
-            new_alpha_vectors.append(alpha)
-        alpha_vectors = jnp.array(new_alpha_vectors)
-    
-    # Compute optimal action
-    values = jnp.dot(alpha_vectors, initial_belief)
-    optimal_action = jnp.argmax(values)
-    
-    return {{
-        "optimal_action": optimal_action,
-        "value": jnp.max(values),
-        "alpha_vectors": alpha_vectors
-    }}
-
-if __name__ == "__main__":
-    # Example usage
-    solver = create_pomdp_solver()
-    print(f"POMDP Solver created with {{solver.num_states}} states, {{solver.num_observations}} observations, {{solver.num_actions}} actions")
-    
-    # Test with uniform initial belief
-    initial_belief = jnp.ones(solver.num_states) / solver.num_states
-    solution = solve_pomdp(solver, initial_belief, horizon=5)
-    
-    print(f"Optimal action: {{solution['optimal_action']}}")
-    print(f"Value: {{solution['value']:.4f}}")
-    print("POMDP solver test successful!")
-'''
-
-        return code
-
-    except Exception as e:
-        raise ValueError(f"JAX POMDP generation failed: {e}") from e
-
-
-def _generate_jax_combined_code(
-    gnn_spec: Dict[str, Any], options: Optional[Dict[str, Any]]
-) -> str:
-    """Generate JAX code for hierarchical/multi-agent/continuous models."""
-
-    model_name = _jax_model_name(gnn_spec, "CombinedModel")
-
-    code = f'''"""
-JAX Combined Model Generated from GNN Specification: {model_name}
-
-This implements a combined JAX model supporting hierarchical, multi-agent, and continuous extensions.
-Uses advanced JAX features including distributed computing and mixed precision.
-
-@Web: https://github.com/google/jax
-@Web: https://optax.readthedocs.io
-@Web: https://flax.readthedocs.io
-"""
-
-import jax
-import jax.numpy as jnp
-import flax.linen as nn
-from functools import partial
-from jax import jit, vmap, pmap
-from typing import Dict, Any, Optional, Tuple, List
-import optax
-
-class {model_name}Combined(nn.Module):
-    """
-    Combined JAX model supporting hierarchical, multi-agent, and continuous extensions.
-    """
-    
-    # Model configuration
-    num_agents: int = 1
-    num_hierarchical_levels: int = 1
-    continuous_dimensions: int = 0
-    use_mixed_precision: bool = True
-    
-    def setup(self):
-        """Initialize model parameters."""
-        # Hierarchical parameters
-        self.hierarchical_weights = []
-        for level in range(self.num_hierarchical_levels):
-            level_weight = self.param(
-                f"hierarchical_{{level}}",
-                nn.initializers.normal(0.1),
-                (self.num_agents, self.num_agents),
-            )
-            self.hierarchical_weights.append(level_weight)
-        
-        # Multi-agent communication parameters
-        if self.num_agents > 1:
-            self.communication_matrix = self.param('communication_matrix',
-                                                 nn.initializers.orthogonal(),
-                                                 (self.num_agents, self.num_agents))
-        
-        # Continuous state parameters
-        if self.continuous_dimensions > 0:
-            self.continuous_encoder = nn.Dense(self.continuous_dimensions)
-            self.continuous_decoder = nn.Dense(self.continuous_dimensions)
-    
-    @nn.compact
-    def __call__(self, inputs: Dict[str, jnp.ndarray], training: bool = False) -> Dict[str, jnp.ndarray]:
-        """
-        Forward pass of the combined model.
-        
-        Args:
-            inputs: Dictionary containing:
-                - 'agent_states': [num_agents, state_dim] - Individual agent states
-                - 'hierarchical_context': [num_levels, context_dim] - Hierarchical context
-                - 'continuous_inputs': [batch_size, continuous_dim] - Continuous inputs
-                - 'communication_mask': [num_agents, num_agents] - Communication topology
-            training: Whether in training mode
-            
-        Returns:
-            Dictionary containing:
-                - 'agent_outputs': [num_agents, output_dim] - Individual agent outputs
-                - 'hierarchical_outputs': [num_levels, output_dim] - Hierarchical outputs
-                - 'continuous_outputs': [batch_size, continuous_dim] - Continuous outputs
-                - 'communication_outputs': [num_agents, num_agents] - Communication outputs
-        """
-        # Extract inputs with defaults
-        agent_states = inputs.get('agent_states', jnp.zeros((self.num_agents, 1)))
-        hierarchical_context = inputs.get('hierarchical_context', jnp.zeros((self.num_hierarchical_levels, 1)))
-        continuous_inputs = inputs.get('continuous_inputs', jnp.zeros((1, self.continuous_dimensions)))
-        communication_mask = inputs.get('communication_mask', jnp.eye(self.num_agents))
-        
-        # The following line generates 'outputs = {{{{}}}}' in the output code
-        outputs = {{}}
-        
-        # 1. Multi-agent processing
-        if self.num_agents > 1:
-            # Apply communication matrix with mask
-            communication_weights = self.communication_matrix * communication_mask
-            
-            # Use vmap to process multiple agents efficiently
-            # jnp.dot(communication_weights, agent_states) is already vectorized, 
-            # but we can explicitly define per-agent processing for complex extensions
-            @jax.vmap
-            def process_agent(weights_row, states):
-                return jnp.dot(weights_row, states)
-                
-            agent_outputs = process_agent(communication_weights, agent_states)
-            outputs['agent_outputs'] = agent_outputs
-            outputs['communication_outputs'] = communication_weights
-        
-        # 2. Hierarchical processing
-        if self.num_hierarchical_levels > 1:
-            hierarchical_outputs = []
-            for level in range(self.num_hierarchical_levels):
-                if level < len(self.hierarchical_weights):
-                    level_output = jnp.dot(self.hierarchical_weights[level], 
-                                         hierarchical_context[level])
-                    hierarchical_outputs.append(level_output)
-                else:
-                    # Default processing for missing levels
-                    hierarchical_outputs.append(hierarchical_context[level])
-            
-            outputs['hierarchical_outputs'] = jnp.stack(hierarchical_outputs)
-        
-        # 3. Continuous processing
-        if self.continuous_dimensions > 0:
-            # Encode continuous inputs
-            encoded = self.continuous_encoder(continuous_inputs)
-            
-            # Apply activation and processing
-            processed = jax.nn.relu(encoded)
-            
-            # Decode back to continuous space
-            decoded = self.continuous_decoder(processed)
-            
-            outputs['continuous_outputs'] = decoded
-        
-        # 4. Combined output (if multiple components exist)
-        if len(outputs) > 1:
-            # Combine different outputs using weighted sum
-            combined_components = []
-            weights = []
-            
-            if 'agent_outputs' in outputs:
-                combined_components.append(outputs['agent_outputs'].flatten())
-                weights.append(1.0)
-            
-            if 'hierarchical_outputs' in outputs:
-                combined_components.append(outputs['hierarchical_outputs'].flatten())
-                weights.append(0.5)
-            
-            if 'continuous_outputs' in outputs:
-                combined_components.append(outputs['continuous_outputs'].flatten())
-                weights.append(0.3)
-            
-            # Normalize weights
-            weights = jnp.array(weights) / jnp.sum(weights)
-            
-            # Weighted combination
-            combined_output = jnp.zeros_like(combined_components[0])
-            for component, weight in zip(combined_components, weights):
-                # Pad or truncate to match size
-                if len(component) > len(combined_output):
-                    component = component[:len(combined_output)]
-                elif len(component) < len(combined_output):
-                    padding = jnp.zeros(len(combined_output) - len(component))
-                    component = jnp.concatenate([component, padding])
-                
-                combined_output += weight * component
-            
-            outputs['combined_output'] = combined_output
-        
-        return outputs
-    
-    def get_parameters(self) -> Dict[str, jnp.ndarray]:
-        """Get all model parameters."""
-        return {
-        "hierarchical_weights": self.hierarchical_weights,
-            'communication_matrix': getattr(self, 'communication_matrix', None),
-            'continuous_encoder': self.continuous_encoder.variables if hasattr(self, 'continuous_encoder') else None,
-            'continuous_decoder': self.continuous_decoder.variables if hasattr(self, 'continuous_decoder') else None
-        }
-    
-    def update_parameters(self, new_params: Dict[str, jnp.ndarray]):
-        """Update model parameters (for training)."""
-        # Implementation for parameter updating during training
-        # This would typically involve gradient-based updates
-        # For now, we'll provide a basic structure
-        updated_params = {{}}
-        
-        for param_name, new_value in new_params.items():
-            if param_name in self.get_parameters():
-                updated_params[param_name] = new_value
-        
-        return updated_params
-
-if __name__ == "__main__":
-    print(f"Combined model {model_name} created successfully!")
-    print("This model supports hierarchical, multi-agent, and continuous extensions.")
-'''
-
-    return code
+from .jax_model_generator import (
+    EFE_CONVENTION_JAX,
+    EFE_CONVENTION_JAX_JSON,
+    EFE_CONVENTION_JAX_JSON_LITERAL,
+    _generate_jax_model_code,
+    _json_dumps,
+)
+from .jax_pomdp_generator import (
+    _generate_jax_pomdp_code,
+)
+from .jax_spec_extract import (
+    _create_fallback_matrix,
+    _create_improved_default_matrix,
+    _extract_gnn_matrices,
+    _infer_matrix_from_context,
+    _jax_model_name,
+    _parse_gnn_matrix_string,
+    _parse_matrix_string,
+    _parse_vector_string,
+    _validated_jax_matrices,
+)
